# FindPacketSenderStake: Remove parallelism to improve performance