#### Problem

The work unit sizes were so small that using the thread pool slowed down this stage significantly.

The existing 1.10 mainnet nodes showed around 30us/packet:
![image](https://user-images.githubusercontent.com/833306/170362804-8f0a87fe-ffc9-486e-ad9b-0363182a37ce.png)

But local tests indicate this should take 10-100ns.

Even when limiting to datapoints with at least 500k packets the processing time per packet stayed above 2us:
![image](https://user-images.githubusercontent.com/833306/170363012-8af28c63-c7ed-47c2-8c92-855c69fe21eb.png)


#### Summary of Changes

Drop the thread pool and `par_iter()`s.

I've tested this on testnet: Before, with 1.10.13:

```
find_packet_sender_stake-services_stats
apply_sender_stakes_time=83549i
total_batches=267i 
total_packets=4902i
```

and with a patched 1.10.19:
```
tpu-vote-find-packet-sender-stake
apply_sender_stakes_time_us=971i
total_batches=299i
total_packets=7752i
```

That means it improved from 17us/packet to 125ns/packet.

@sakridge